+++ src/hwenc_jetson/jetson_video_encoder.cpp
@@ -672,7 +672,6 @@ webrtc::VideoEncoder::EncoderInfo JetsonVideoEncoder::GetEncoderInfo() const {
     info.scaling_settings =
         VideoEncoder::ScalingSettings(kLowVp9QpThreshold, kHighVp9QpThreshold);
   }
-  info.has_internal_source = false;
   return info;
 }